First commit. Added a mysql stack.
This is a simple mysql stack. It uses a persistent volume to store the data. It also uses a secret to store the root password. The secret is created by the stack. The secret is used by the mysql container to set the root password. The secret is also used by the mysql client container to connect to the mysql server.